# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SocialApp API",
        "description": "API for SocialApp",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/friends": {
            "get": {
                "tags": [
                    "Friendships"
                ],
                "summary": "List all friends with pagination",
                "operationId": "b749de143f5470e2f8013e89de1ac3f8",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of friends with pagination",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "friends": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Friendship"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Create a new friendship request",
                "operationId": "e458663142413fd7ed044a08ed383c65",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "friend_id"
                                ],
                                "properties": {
                                    "friend_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Friendship request created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Friendship"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "A friendship request already exists or is pending between these users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "A friendship request already exists or is pending between these users."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends/pending": {
            "get": {
                "tags": [
                    "Friendships"
                ],
                "summary": "List all pending friendship requests",
                "operationId": "4a7fe6223c5ffb16f9f5a7e647a01a85",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of pending friendship requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pendingRequests": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Friendship"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total_pages": {
                                                    "type": "integer"
                                                },
                                                "total_items": {
                                                    "type": "integer"
                                                },
                                                "has_more_pages": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends/{friendship}": {
            "put": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Update the status of a friendship",
                "operationId": "c8d4a5f038763b29ee6afd87e7954965",
                "parameters": [
                    {
                        "name": "friendship",
                        "in": "path",
                        "description": "The ID of the friendship to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "accepted"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Friendship status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Friendship"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Delete a friendship",
                "operationId": "bcd88d291d7db59acdda042eaac2d656",
                "parameters": [
                    {
                        "name": "friendship",
                        "in": "path",
                        "description": "The ID of the friendship to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friendship deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Friendship deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Friendship": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "friend_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "Pagination": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "total_pages": {
                        "type": "integer",
                        "example": 15
                    },
                    "total_items": {
                        "type": "integer",
                        "example": 150
                    },
                    "has_more_pages": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token obtained from logging in",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}